backend/requirements.server.txt
fastapi>=0.83.0
uvicorn[standard]>=0.18.0
python-multipart>=0.0.5
pydantic>=2.0.0
pydantic-settings>=2.0.0
openpyxl>=3.0.0
pandas>=2.0.0
PyJWT>=2.12.0
pwdlib[bcrypt]>=0.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
